requirements.txt
dominate==2.3.1
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.3
Flask-Injector==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
idna==2.6
inflection==0.3.1
injector==0.12.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
marshmallow==3.0.0b7
marshmallow-sqlalchemy==0.13.2
mccabe==0.6.1
pycparser==2.18
PyJWT==1.6.1
python-dateutil==2.7.2
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.4
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.5
typing==3.6.2
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
